requirements.txt
websockets==11.0.3
ollama==0.4.7
opencv-python==4.8.1.78
SpeechRecognition==3.14.1
pyttsx3==2.90
numpy==1.24.3
pipwin==0.5.2
pyaudio==0.2.14
fastapi==0.109.2
uvicorn==0.27.1
Jinja2==3.1.3
Pillow==10.2.0
aiofiles==23.2.1